+++ common/djangoapps/student/management/commands/create_random_users.py
@@ -9,7 +9,7 @@ from django.core.management.base import BaseCommand
 from opaque_keys.edx.keys import CourseKey
 from six.moves import range
 
-from student.forms import AccountCreationForm
+from openedx.core.djangoapps.user_authn.views.registration_form import AccountCreationForm
 from student.helpers import do_create_account
 from student.models import CourseEnrollment
